pyproject.toml
[project]
name = "workshop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ape-vyper>=0.8.9",
    "eth-ape>=0.8.27",
    "ipython>=8.32.0",
    "snekmate>=0.1.0",
    "titanoboa>=0.2.5",
    "vyper>=0.4.0",
]
